+++ devicelayer/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+- Device Configuration logs.
 ### Security
 
 ## [0.3.0] - 2022-6-17